[REFACTOR] issueDetail 컴포넌트 모듈화
1. issueHeader
2. issueBody (issueContent, issueSideBar)
