+++ client/src/components/issueDetail/issueDetailHeader.js
@@ -1,37 +1,37 @@
-// import React, { useState } from 'react';
-// import styled from 'styled-components';
-// import { useIssues } from '../issuelist/issueHook.js';
+import React, { useState } from 'react';
+import styled from 'styled-components';
 
-// const IssueHeadContainer = styled.div`
-//   border-bottom: 1px solid #d1d5da;
-//   margin-bottom: 20px;
-// `;
+const IssueHeadContainer = styled.div`
+  border-bottom: 1px solid #d1d5da;
+  margin-bottom: 20px;
+`;
 
-// const EditTitleButton = styled.button`
-//   float: right;
-//   padding: 10px;
-//   margin-right: 100px;
-// `;
+const EditTitleButton = styled.button`
+  float: right;
+  padding: 10px;
+  margin-right: 100px;
+`;
 
-// const Header = ({ match }) => {
-//     const editTitle = (e) => {
-//         setTitle(e.target.value);
-//     };
+const SaveTitleButton = styled.button``;
 
-//     const issues = useIssues();
-//     const { id } = match.params;
-//     console.log(id);
-//     const issue = issues[id - 1];
+const CancelTitleButton = styled.button``;
 
-//     return (
-//         <>
-//             <IssueHeadContainer>
-//                 <EditTitleButton onClick={editTitle}>Edit</EditTitleButton>
-//                 <h3>{issue.title} #{issue.id}</h3>
-//                 {issue.is_open === 1 ? <p>Open</p> : <p>Closed</p>}
-//             </IssueHeadContainer>
-//         </>
-//     )
-// }
+const EditContentsButton = styled.button``;
 
-// export default Header;
+const IssueHeader = (props) => {
+    const editTitle = (e) => {
+        setTitle(e.target.value);
+    };
+    
+    return (
+        <>
+            <IssueHeadContainer>
+                <EditTitleButton onClick={editTitle}>Edit</EditTitleButton>
+                <h3>{props.title} #{props.id}</h3>
+                {props.is_open === 1 ? <p>Open</p> : <p>Closed</p>}
+            </IssueHeadContainer>
+        </>
+    )
+}
+
+export default IssueHeader;
